Moved ytm implementations from Base to YtmRetriever where it belongs
